feat: moved raw api calls to domain level files
