+++ src/components/event/mobile/TopStakesSectionMobile.tsx
@@ -2,7 +2,7 @@ import Avatar from "@mui/material/Avatar";
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import type { EventStake } from "../../../api";
+import type { EventStake } from "../../../api/api";
 import type { Pool } from "../../../models/models";
 
 export function TopStakesSectionMobile({
