+++ src/hooks/queries/useEventsQuery.ts
@@ -1,7 +1,7 @@
 import { useAuth } from "@clerk/react";
 import { useQuery } from "@tanstack/react-query";
 import type { UUID } from "../../models/models";
-import { fetchEvents } from "../../api";
+import { fetchEvents } from "../../api/api";
 import { queryKeys } from "../queryKeys";
 
 export function useEventsQuery(topicId: UUID) {
