+++ src/hooks/queries/useLeaderboardQuery.ts
@@ -1,7 +1,7 @@
 import { useAuth } from "@clerk/react";
 import { useQuery, type UseQueryResult } from "@tanstack/react-query";
 import { queryKeys } from "../queryKeys";
-import { fetchLeaderboard, type LeaderboardEntry } from "../../api";
+import { fetchLeaderboard, type LeaderboardEntry } from "../../api/api";
 
 export function useLeaderboardQuery(
   limit: number,
