+++ src/hooks/queries/usePoolsQuery.ts
@@ -1,6 +1,6 @@
 import { useAuth } from "@clerk/react";
 import { useQuery } from "@tanstack/react-query";
-import { fetchPools } from "../../api";
+import { fetchPools } from "../../api/api";
 import type { UUID } from "../../models/models";
 import { queryKeys } from "../queryKeys";
