+++ src/hooks/queries/usePortfolioQuery.ts
@@ -1,6 +1,6 @@
 import { useAuth } from "@clerk/react";
 import { useQuery } from "@tanstack/react-query";
-import { fetchPortfolio } from "../../api";
+import { fetchPortfolio } from "../../api/api";
 import { queryKeys } from "../queryKeys";
 
 export function usePortfolioQuery() {
