+++ src/hooks/queries/useTopicEventCountQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import type { UUID } from "../../models/models";
-import { fetchTopicEventCounts } from "../../api";
+import { fetchTopicEventCounts } from "../../api/api";
 import { queryKeys } from "../queryKeys";
 
 export function useTopicEventCountQuery(topicId: UUID) {
