+++ src/views/UserView.tsx
@@ -6,7 +6,7 @@ import Box from "@mui/material/Box";
 import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { fetchUserPortfolio } from "../api";
+import { fetchUserPortfolio } from "../api/api";
 import type { UUID } from "../models/models";
 import { PortfolioStakeSections } from "../components/portfolio/PortfolioStakeSections";
